feat: Added prop-types to help with potential debugging when passing props. Setup basic page HTML structure + created stylesheet object
At this point have got a good starting point for the project. Connected a stylesheet to my component and interestingly enough it worked without needing the StyleSheet API that comes packaged in the react-native package (but does not come in the base react library). This is just a plain JS object. Tested changing background color to prove that I can now grab hold of HTML elements and change their styling. That worked and now I am in business. Also created a placeholder handleClick() method to trigger the mining