feat: Created utils helper file with SHA256 algorithm and hooked it up to UI + improved overall styling
Learnt about the differences between <input> and <textarea> elements. Note that there are a few console warnings that I should deal with in near future. Also how I miss going back to devving for the Web. You got at your easy disposal all the dev tools and chrome extensions for React whereas for mobile it was so hard