feat: Fixed and brushed up the string concatenation feature as input into sha256 algorithm
Learnt that this.setstate() is actually asynchronous and I had to use some callback function to get around this. Didn't know this at first but I started to realize it when I saw the weird behavior in my React Chrome Devtools. this.state.concatenatedString was always one step behind and also the output of the sha256 function was always one step behind