feat: Hooked up Redux to project. Designed data structure. Created actions and reducers
Also hooked it up to Chrome Dev Tools and checked and the store is appearing containing the global app state! Initially it was just two blank strings as they were just being initialized. Later on I loaded the data in from the mock DB which is just an in-memory JS object and then saved it to the Redux store and now I can see all my data there! Redux is confusing as usual but this time took me much shorter than previous times and it is starting to make a little more sense. At least I could figure it out myself without assistance