Subsequent block MVP functionality
Made a design decision to repeat some code and copied and pasted the bulk of GenesisBlock code into SubsequentBlock because things were getting far too confusing. So I forgoed elegance and brevity of code for simplicity. In this case it was different from previous projects, though the two components were pretty similar they were still fundamentally pretty different and it was getting way too complicated. Once I took it out it greatly simplified the rest of the work. I also learnt you don't have to map through a list you can just map through a React component but then the problem is I don't think it starts a new line like a list would + I created a helper function to generate a UUID for the purpose of the mapping function